solved some problems with sphinx in rst syntax: cant call to autoclass in text when autosummary is here without some protection (maybe :noindex:?), also stars from **kwargs need to be escaped and further details
