+++ coopihc/examples/simple_examples/inference_examples.py
@@ -1,0 +1,10 @@
+import sys
+from pathlib import Path
+
+file = Path(__file__).resolve()
+root = file.parents[3]
+sys.path.append(str(root))
+
+from coopihc.inference.ExampleInferenceEngine import ExampleInferenceEngine
+
+eie = ExampleInferenceEngine()
