+++ coopihc/examples/simple_examples/interactiontask_examples.py
@@ -1,0 +1,10 @@
+import sys
+from pathlib import Path
+
+file = Path(__file__).resolve()
+root = file.parents[3]
+sys.path.append(str(root))
+
+from coopihc.interactiontask.ExampleTask import ExampleTask
+
+et = ExampleTask()
