+++ coopihc/examples/simple_examples/stateelement_examples.py
@@ -1,3 +1,10 @@
+import sys
+from pathlib import Path
+
+file = Path(__file__).resolve()
+root = file.parents[3]
+sys.path.append(str(root))
+
 if __name__ == "__main__":
 
     import numpy
