+++ coopihc/examples/worked_out_examples/modules.py
