+++ coopihc/examples/worked_out_examples/train-and-execute.py
