+++ coopihc/inference/ExampleInferenceEngine.py
@@ -1,10 +1,5 @@
 from coopihc.inference.BaseInferenceEngine import BaseInferenceEngine
 
-
 class ExampleInferenceEngine(BaseInferenceEngine):
     def infer(self):
         return self.state, 0
-
-
-if __name__ == "__main__":
-    eie = ExampleInferenceEngine()
