+++ docs/Makefile
@@ -13,6 +13,13 @@ help:
 
 .PHONY: help Makefile
 
+whtml:
+	
+	rm -rf _build
+	rm -rf _autosummary
+	@$(SPHINXBUILD) -M html "$(SOURCEDIR)" "$(BUILDDIR)" -W $(SPHINXOPTS) $(O)
+
+
 # Catch-all target: route all unknown targets to Sphinx using the new
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
